requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiosqlite==0.19.0
astunparse==1.6.3
asyncua==1.0.3
cachetools==5.3.1
casadi==3.6.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudpickle==2.2.1
contourpy==1.1.0
control==0.9.4
cryptography==41.0.3
cycler==0.11.0
do-mpc==4.6.0
flatbuffers==23.5.26
fonttools==4.42.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
gym==0.26.2
gym-notices==0.0.8
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
onnx==1.14.0
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
polytope==0.2.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycddlib==2.1.6
pycparser==2.21
pygame==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sortedcontainers==2.4.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
typing_extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
